+++ lib/r18gs/index.ts
@@ -1,5 +1,0 @@
-import useRGS from "./src";
-
-export default useRGS;
-
-export * from "./src";