+++ lib/r18gs/tests/use-rgs.test.tsx
@@ -4,6 +4,8 @@ import useRGS from "../src";
 import { ChangeEvent, useCallback } from "react";
 
 const COUNT_RGS_KEY = "count";
+const TESTID_INPUT = "in1";
+const TESTID_DISPLAY = "d1";
 
 function Component1() {
 	const [count, setCount] = useRGS<number>(COUNT_RGS_KEY, 0);
@@ -15,21 +17,21 @@ function Component1() {
 	);
 	return (
 		<div>
-			<input data-testid="input" onChange={handleChange} type="number" value={count} />
+			<input data-testid={TESTID_INPUT} onChange={handleChange} type="number" value={count} />
 		</div>
 	);
 }
 
 function Component2() {
 	const [count] = useRGS<number>(COUNT_RGS_KEY);
-	return <h1 data-testid="display">{count}</h1>;
+	return <h1 data-testid={TESTID_DISPLAY}>{count}</h1>;
 }
 
 describe("React18GlobalStore", () => {
 	test("check state update to multiple components", async ({ expect }) => {
 		render(<Component1 />);
 		render(<Component2 />);
-		await act(() => fireEvent.input(screen.getByTestId("input"), { target: { value: 5 } }));
-		expect(screen.getByTestId("display").textContent).toBe("5");
+		await act(() => fireEvent.input(screen.getByTestId(TESTID_INPUT), { target: { value: 5 } }));
+		expect(screen.getByTestId(TESTID_DISPLAY).textContent).toBe("5");
 	});
 });