+++ lib/r18gs/vitest.config.ts
@@ -11,7 +11,6 @@ export default defineConfig({
 		setupFiles: [],
 		coverage: {
 			include: ["src/**"],
-			exclude: ["src/**/index.ts", "src/**/declaration.d.ts"],
 			reporter: ["text", "json", "clover", "html"],
 		},
 	},